+++ packages/web3-eth-transaction-utils/src/constants.ts
@@ -1,4 +1,4 @@
-﻿/*
+/*
 This file is part of web3.js.
 
 web3.js is free software: you can redistribute it and/or modify
@@ -14,19 +14,6 @@ GNU Lesser General Public License for more details.
 You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
+import { FMT_BYTES, FMT_NUMBER } from 'web3-utils';
 
-export * from './error_types';
-export * from './apis/eth_execution_api';
-export * from './apis/web3_eth_execution_api';
-export * from './apis/web3_net_api';
-export * from './apis/eth_personal_api';
-export * from './eth_types';
-export * from './json_rpc_types';
-export * from './primitives_types';
-export * from './utility_types';
-export * from './web3_api_types';
-export * from './web3_base_provider';
-export * from './web3_base_wallet';
-export * from './web3_deferred_promise_type';
-export * from './private_key_to_address';
-export * from './web3_eth_schemas';
+export const NUMBER_DATA_FORMAT = { bytes: FMT_BYTES.HEX, number: FMT_NUMBER.NUMBER } as const;