+++ packages/web3-eth/test/unit/rpc_method_wrappers/send_signed_transaction.test.ts
@@ -30,7 +30,6 @@ import {
 import { transactionReceiptSchema } from '../../../src/schemas';
 
 jest.mock('web3-rpc-methods');
-jest.mock('../../../src/utils/wait_for_transaction_receipt');
 jest.mock('../../../src/utils/watch_transaction_for_confirmations');
 
 describe('sendTransaction', () => {
@@ -48,9 +47,9 @@ describe('sendTransaction', () => {
 	it.each(testData)(
 		`sending event should emit with inputSignedTransaction\n ${testMessage}`,
 		async (_, inputSignedTransaction) => {
-			(
-				WaitForTransactionReceipt.waitForTransactionReceipt as jest.Mock
-			).mockResolvedValueOnce(expectedTransactionReceipt);
+			(ethRpcMethods.getTransactionReceipt as jest.Mock).mockResolvedValueOnce(
+				expectedTransactionReceipt,
+			);
 
 			const inputSignedTransactionFormatted = format(
 				{ eth: 'bytes' },
@@ -72,9 +71,9 @@ describe('sendTransaction', () => {
 	it.each(testData)(
 		`should call ethRpcMethods.sendRawTransaction with expected parameters\n ${testMessage}`,
 		async (_, inputSignedTransaction) => {
-			(
-				WaitForTransactionReceipt.waitForTransactionReceipt as jest.Mock
-			).mockResolvedValueOnce(expectedTransactionReceipt);
+			(ethRpcMethods.getTransactionReceipt as jest.Mock).mockResolvedValueOnce(
+				expectedTransactionReceipt,
+			);
 
 			const inputSignedTransactionFormatted = format(
 				{ eth: 'bytes' },
@@ -92,9 +91,9 @@ describe('sendTransaction', () => {
 	it.each(testData)(
 		`sent event should emit with inputSignedTransaction\n ${testMessage}`,
 		async (_, inputSignedTransaction) => {
-			(
-				WaitForTransactionReceipt.waitForTransactionReceipt as jest.Mock
-			).mockResolvedValueOnce(expectedTransactionReceipt);
+			(ethRpcMethods.getTransactionReceipt as jest.Mock).mockResolvedValueOnce(
+				expectedTransactionReceipt,
+			);
 
 			const inputSignedTransactionFormatted = format(
 				{ eth: 'bytes' },
@@ -117,9 +116,9 @@ describe('sendTransaction', () => {
 	it.each(testData)(
 		`transactionHash event should emit with inputSignedTransaction\n ${testMessage}`,
 		async (_, inputSignedTransaction) => {
-			(
-				WaitForTransactionReceipt.waitForTransactionReceipt as jest.Mock
-			).mockResolvedValueOnce(expectedTransactionReceipt);
+			(ethRpcMethods.getTransactionReceipt as jest.Mock).mockResolvedValueOnce(
+				expectedTransactionReceipt,
+			);
 
 			(ethRpcMethods.sendRawTransaction as jest.Mock).mockResolvedValueOnce(
 				expectedTransactionHash,
@@ -138,18 +137,18 @@ describe('sendTransaction', () => {
 	);
 
 	it.each(testData)(
-		`should call WaitForTransactionReceipt.waitForTransactionReceipt with expected parameters\n ${testMessage}`,
+		`should call ethRpcMethods.getTransactionReceipt with expected parameters\n ${testMessage}`,
 		async (_, inputSignedTransaction) => {
 			(ethRpcMethods.sendRawTransaction as jest.Mock).mockResolvedValueOnce(
 				expectedTransactionHash,
 			);
-			(
-				WaitForTransactionReceipt.waitForTransactionReceipt as jest.Mock
-			).mockResolvedValueOnce(expectedTransactionReceipt);
+			(ethRpcMethods.getTransactionReceipt as jest.Mock).mockResolvedValueOnce(
+				expectedTransactionReceipt,
+			);
 
 			await sendSignedTransaction(web3Context, inputSignedTransaction, DEFAULT_RETURN_FORMAT);
-			expect(WaitForTransactionReceipt.waitForTransactionReceipt).toHaveBeenCalledWith(
-				web3Context,
+			expect(ethRpcMethods.getTransactionReceipt).toHaveBeenCalledWith(
+				web3Context.requestManager,
 				expectedTransactionHash,
 				DEFAULT_RETURN_FORMAT,
 			);
@@ -186,9 +185,9 @@ describe('sendTransaction', () => {
 				DEFAULT_RETURN_FORMAT,
 			);
 
-			(
-				WaitForTransactionReceipt.waitForTransactionReceipt as jest.Mock
-			).mockResolvedValueOnce(expectedTransactionReceipt);
+			(ethRpcMethods.getTransactionReceipt as jest.Mock).mockResolvedValueOnce(
+				expectedTransactionReceipt,
+			);
 
 			await sendSignedTransaction(
 				web3Context,
@@ -211,9 +210,9 @@ describe('sendTransaction', () => {
 				DEFAULT_RETURN_FORMAT,
 			);
 
-			(
-				WaitForTransactionReceipt.waitForTransactionReceipt as jest.Mock
-			).mockResolvedValueOnce(expectedTransactionReceipt);
+			(ethRpcMethods.getTransactionReceipt as jest.Mock).mockResolvedValueOnce(
+				expectedTransactionReceipt,
+			);
 			expect(
 				await sendSignedTransaction(
 					web3Context,
@@ -240,9 +239,9 @@ describe('sendTransaction', () => {
 			(ethRpcMethods.sendRawTransaction as jest.Mock).mockResolvedValueOnce(
 				expectedTransactionHash,
 			);
-			(
-				WaitForTransactionReceipt.waitForTransactionReceipt as jest.Mock
-			).mockResolvedValueOnce(expectedTransactionReceipt);
+			(ethRpcMethods.getTransactionReceipt as jest.Mock).mockResolvedValueOnce(
+				expectedTransactionReceipt,
+			);
 
 			const promiEvent = sendSignedTransaction(
 				web3Context,
@@ -252,8 +251,8 @@ describe('sendTransaction', () => {
 
 			await promiEvent;
 
-			expect(WaitForTransactionReceipt.waitForTransactionReceipt).toHaveBeenCalledWith(
-				web3Context,
+			expect(ethRpcMethods.getTransactionReceipt).toHaveBeenCalledWith(
+				web3Context.requestManager,
 				expectedTransactionHash,
 				DEFAULT_RETURN_FORMAT,
 			);